requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
backcall==0.1.0
billiard==3.6.0.0
boto3==1.9.188
botocore==1.12.215
cairocffi==1.0.2
CairoSVG==2.4.0
celery==4.3.0
certifi==2019.3.9
cffi==1.11.5
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect2==0.2.1
daphne==2.2.3
decorator==4.3.0
defusedxml==0.6.0
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-reverse-admin==2.3.0
django-scheduler==0.8.8
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-gis==0.13
djangorestframework-jwt==1.11.0
docutils==0.14
factory-boy==2.11.1
Faker==1.0.0
hiredis==0.3.0
html5lib==1.0.1
hyperlink==18.0.0
icalendar==4.0.3
idna==2.7
incremental==17.5.0
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
jmespath==0.9.3
kombu==4.5.0
msgpack==0.5.6
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.0
PyHamcrest==1.9.0
PyJWT==1.6.4
Pyphen==0.9.5
python-dateutil==2.7.5
pytz==2018.7
redis==3.2.1
requests==2.22.0
s3transfer==0.2.1
six==1.11.0
text-unidecode==1.2
tinycss2==1.0.2
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
WeasyPrint==47
webencodings==0.5.1
xmltodict==0.12.0
django-request-logging
django-sysinfo
